% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_Kset}
\alias{build_Kset}
\title{Build K_set}
\usage{
build_Kset(Nset, Basis, n, print_bool = FALSE)
}
\arguments{
\item{Nset}{A list containing 2 vectors which correspond to a index based mapping of the states in the data set to their occurrence.}

\item{Basis}{The basis on which the model is based.}

\item{n}{Number of binary (spin) variables of the data set.}

\item{print_bool}{print relevant information. Default: FALSE.}
}
\value{
Kset Mapping of frequency of occurrence of each state of the data-set in the new basis.
}
\description{
Build Kset for the states written in the basis of the m-chosen independent 
operator on which the SC model is based. Changes the basis of the data-set from its original basis
to the basis provided as an argument in basis. Print Kset to terminal by setting print_bool to TRUE.
}
