% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_partition}
\alias{check_partition}
\title{check_partition}
\usage{
check_partition(Partition, n)
}
\arguments{
\item{Partition}{Partition of MCM.}

\item{n}{The amount of binary (spin) variables.}
}
\value{
Bool_results Returns FALSE if there is an overlap between the parts.
}
\description{
Check if *Partition* is an actual partition of r basis elements,
i.e., that no basis element appears in more than 1 part of the partition.
i.e., that each basis element only appears in a single part of the partition.
}
