% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Complexity_MCM}
\alias{Complexity_MCM}
\title{Complexity of an MCM based on the geometric and parametric complexity.}
\usage{
Complexity_MCM(Partition, N, n, C_param, C_geom)
}
\arguments{
\item{Partition}{Partition of MCM.}

\item{N}{The size of the dataset}

\item{n}{The amount of spin variables.}

\item{C_param}{The (parametric) first order complexity.}

\item{C_geom}{The geometric complexity.}
}
\value{
Complexity A double containing the complexity of the MCM.
}
\description{
Complexity of and MCM defined by the addition of the total geometric complexity and
total (parametric) first order complexity. 
Compute separately: -- the first order complexity    --> stored in C_param
                    -- and the geometric complexity  --> stored in C_geom
}
