% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Create_MCM}
\alias{Create_MCM}
\title{Define an MCM}
\usage{
Create_MCM(MCM_table)
}
\arguments{
\item{MCM_table}{List containing the integer representation of the MCM parts.}
}
\value{
MCM_partition The defined MCM.
}
\description{
Define an MCM. Input is an integer representation of all MCM parts. For example, 
MCM_Choice = c(384, 64, 32, 16, 8, 4, 2, 1) defines an MCM with 8 independent parts, 
based on n=9 spins. You can check that the model (i.e., list of parts) that you have 
provided properly defines an MCM by calling the function check_partition(Partition, n).
}
