% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GeomComplexity_SubCM}
\alias{GeomComplexity_SubCM}
\title{GeomComplexity_SubCM}
\usage{
GeomComplexity_SubCM(m)
}
\arguments{
\item{m}{The number of basis elements/operators.}
}
\value{
GeomComplexity The complexity.
}
\description{
Geometric Complexity of a Sub-Complete Model based on m basis operators.
for 1 <= m <= n . Rem C(m=1) = log(pi)
}
