% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogE_MCM}
\alias{LogE_MCM}
\title{Compute the log-evidence of of an MCM.}
\usage{
LogE_MCM(Kset, Partition, N, n, print_bool = FALSE)
}
\arguments{
\item{Kset}{Mapping of frequency of occurrence of each state of the data-set in the new basis.}

\item{Partition}{Partition of MCM.}

\item{N}{The size of the data-set.}

\item{n}{The amount of binary (spin) variables.}

\item{print_bool}{Boolean to print related information to terminal. Default: FALSE.}
}
\value{
LogE The log-evidence of an MCM as a double.
}
\description{
Compute the log-evidence of of an MCM.
}
