% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogE_SubCM}
\alias{LogE_SubCM}
\title{Compute the log-evidence of the sub-complete part (of an MCM).}
\usage{
LogE_SubCM(Kset, Ai, N, n, print_bool = FALSE)
}
\arguments{
\item{Kset}{Mapping of frequency of occurrence of each state of the data-set in the new basis.}

\item{Ai}{The binary representation of the sub-complete part (of an MCM).}

\item{N}{The size of the data-set.}

\item{n}{The amount of binary (spin) variables.}

\item{print_bool}{Boolean to print related information to terminal. Default: FALSE.}
}
\value{
LogE The log-evidence of the sub-complete part (of an MCM) as a double.
}
\description{
Compute the log-evidence of the sub-complete part (of an MCM) defined by Ai.
This function could be also used directly by the user
to compute the log-likelihood of a sub-complete model
}
