% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogL_CM}
\alias{LogL_CM}
\title{Compute the log-likelihood of a complete model on Kset.}
\usage{
LogL_CM(Kset, N)
}
\arguments{
\item{Kset}{Mapping of frequency of occurrence of each state of the data-set in the new basis.}

\item{N}{The size of the data-set.}
}
\value{
LogL The log-likelihood of a complete model on Kset as a double.
}
\description{
Compute the log-likelihood of a complete model on Kset:
This function is mainly used for call by `LogL_SC_PartMCM`,
but can also be used to compute the log-likelihood of a complete model
}
