% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogL_MCM}
\alias{LogL_MCM}
\title{Compute log-likelihood of an MCM.}
\usage{
LogL_MCM(Kset, Partition, N, n, print_bool = FALSE)
}
\arguments{
\item{Kset}{Mapping of frequency of occurrence of each state of the data-set in the new basis.}

\item{Partition}{Partition of MCM.}

\item{N}{The size of the data-set.}

\item{n}{Number of binary (spin) variables of the data set.}

\item{print_bool}{Boolean to print related information to terminal. Default: FALSE.}
}
\value{
LogL The log-likelihood of a complete model on Kset as a double.
}
\description{
Compute the log-likelihood of of an MCM.
This function could be also used directly by the user
to compute the log-likelihood of a model.
}
