% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogL_SubCM}
\alias{LogL_SubCM}
\title{Compute log-likelihood of the sub-complete part (SC-part) of an MCM.}
\usage{
LogL_SubCM(Kset, Ai, N, n, print_bool = FALSE)
}
\arguments{
\item{Kset}{Mapping of frequency of occurrence of each state of the data-set in the new basis.}

\item{Ai}{The binary representation of the SC-part.}

\item{N}{The size of the data-set.}

\item{n}{Number of binary (spin) variables of the data set.}

\item{print_bool}{Boolean to print related information to terminal. Default: FALSE.}
}
\value{
LogL The log-likelihood of a sub-complete part (of an MCM) defined by Ai.
}
\description{
Compute the log-likelihood of the sub-complete part (of an MCM) defined by Ai.
This function could be also used directly by the user to compute the 
log-likelihood of a sub-complete model
}
