% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MCM_GivenRank_r}
\alias{MCM_GivenRank_r}
\title{MCM_GivenRank_r}
\usage{
MCM_GivenRank_r(Kset, N, LogE_best, r, n, OUTPUT_directory, print_bool)
}
\arguments{
\item{Kset}{Mapping of frequency of occurrence of each state of the data-set in the new basis.}

\item{N}{The size of the data-set.}

\item{LogE_best}{The double which will contain the maximized Log-evidence.}

\item{r}{The rank r which is the r first elements of the basis.}

\item{n}{The amount of binary (spin) variables.}

\item{OUTPUT_directory}{The name of the output directory (and path to it) in which to save the output data.}

\item{print_bool}{Boolean to print related information to terminal. Default: FALSE.}
}
\value{
Partition The best MCM.
}
\description{
Compare all the MCM of rank r,
based on the r first elements of the basis used to build Kset:
}
