% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MCM_search}
\alias{MCM_search}
\title{Perform the exhaustive search for the best MCM at once.}
\usage{
MCM_search(n, datafilename, Basis_li, OUTPUT_directory, comparison)
}
\arguments{
\item{n}{Number of binary (spin) variables of the data set.}

\item{datafilename}{The name of input file including path containing the binary data.}

\item{Basis_li}{The basis on which the model is based.}

\item{OUTPUT_directory}{The name of the output directory (and path to it) in which to save the output data.}

\item{comparison}{Boolean indicating whether to run 2 other versions of the exhaustive search. For comparison purposes. Default: FALSE.}
}
\description{
Performs the community detection on the data given in datafilename with the given basis Basis_li 
and returns the best fitting MCM for the data. Optional: set the comparison
variable to TRUE to test what happens when the model is compared to 3 different cases.
For general use set to FAlSE. Output is stored in the (newly created)
OUTPUT directory in the same working directory.
}
