% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Original_Basis}
\alias{Original_Basis}
\title{Original Basis}
\usage{
Original_Basis(n)
}
\arguments{
\item{n}{Number of binary (spin) variables of the data set.}
}
\value{
Basis_li List of the original basis i.e. {s1, s2, ..., sn}.
}
\description{
Return the original basis, i.e., {s1, s2, ..., sn}
}
