% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ParamComplexity_SubCM}
\alias{ParamComplexity_SubCM}
\title{Parametric Complexity of a sub-complete model}
\usage{
ParamComplexity_SubCM(m, N)
}
\arguments{
\item{m}{The number of basis elements/operators.}

\item{N}{The size of the data-set.}
}
\value{
ParamComplexity The complexity.
}
\description{
Parametric Complexity of a Sub-Complete Model based on m basis operators.
for 1 <= m <= n . Rem C(m=1) = log(pi).
}
