% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrintInfo_All_Indep_Models}
\alias{PrintInfo_All_Indep_Models}
\title{PRINT INFO ON SUCCESSIVE INDEPENDENT MODELS IN THE NEW BASIS.}
\usage{
PrintInfo_All_Indep_Models(Kset, N, n)
}
\arguments{
\item{Kset}{Mapping of frequency of occurrence of each state of the data-set in the new basis.}

\item{N}{The size of the data-set.}

\item{n}{The amount of binary (spin) variables.}
}
\description{
The function prints the integer and binary representation of the part 
, as well as the the log-likelihood (LogL) and the log-evidence (LogE).
}
