% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrintTerm_Basis}
\alias{PrintTerm_Basis}
\title{Print Basis}
\usage{
PrintTerm_Basis(Basis_li, n)
}
\arguments{
\item{Basis_li}{A list containing the basis.}

\item{n}{Number of binary (spin) variables of the data set.}
}
\description{
Print the basis info in the terminal.
}
