% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrintTerminal_MCM_Info}
\alias{PrintTerminal_MCM_Info}
\title{PRINT INFO on each PART of an MCM (= a partition)}
\usage{
PrintTerminal_MCM_Info(Kset, N, n, MCM_Partition)
}
\arguments{
\item{Kset}{Mapping of frequency of occurrence of each state of the data-set in the new basis.}

\item{N}{The size of the data-set.}

\item{n}{The amount of binary (spin) variables.}

\item{MCM_Partition}{Partition of MCM.}
}
\description{
The function prints the number of Independent Components of the MCM 
(i.e., the number of partitions -- or communities -- of the MCM),
as well as the log-likelihood (LogL), the parameter complexity (C_param), 
the geometric complexity (C_geom), the total complexity (C_tot), 
the Minimum Description Length (MDL) and the log-evidence (LogE).
}
