% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Read_BasisOp_BinaryRepresentation}
\alias{Read_BasisOp_BinaryRepresentation}
\title{Read basis operators in binary representation from input file.}
\usage{
Read_BasisOp_BinaryRepresentation(Basis_binary_filename, n)
}
\arguments{
\item{Basis_binary_filename}{The name of input file including path.}

\item{n}{Number of binary (spin) variables of the data set.}
}
\value{
Basis_li List of the basis with the integer value of that binary representation.
}
\description{
Reads the Basis operators in binary representation from the given input 
file. The operators should be written in one single column and as 
binary representation of the spin involved in the Operator. Returns a 
list of the basis with the integer value of that binary representation
}
