% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Read_BasisOp_IntegerRepresentation}
\alias{Read_BasisOp_IntegerRepresentation}
\title{Read basis operators in integer representation from input file.}
\usage{
Read_BasisOp_IntegerRepresentation(Basis_integer_filename)
}
\arguments{
\item{Basis_integer_filename}{The name of input file including path.}
}
\value{
Basis_li List of the given input file.
}
\description{
Reads the Basis operators in integer representation from the given input 
file. The operators should be written in one single column and as the
integer value of that binary representation. Returns a 
list of the given input file.
}
