% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_datafile}
\alias{read_datafile}
\title{Read data and store in Nset.}
\usage{
read_datafile(N, n, filename)
}
\arguments{
\item{N}{Integer that will contain the data set size.}

\item{n}{Number of binary (spin) variables of the data set.}

\item{filename}{The string containing the path to the data file and the name.}
}
\value{
Nset The first row contains the observed state and the second row contains their occurrence (matching column-wise).
}
\description{
Reads the binary data set (with location and name provided in argument filename) 
as binary integers and returns a mapping of each observed state to the number of times 
they occur in the data set. Note that each state of the system is encoded as an n-bit integer.
}
