% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Read_MCMParts_BinaryRepresentation}
\alias{Read_MCMParts_BinaryRepresentation}
\title{Read_MCMParts_BinaryRepresentation}
\usage{
Read_MCMParts_BinaryRepresentation(MCM_binary_filename, n)
}
\arguments{
\item{MCM_binary_filename}{The file name (incl. the path to it from the current directory).}

\item{n}{The amount of binary (spin) variables.}
}
\value{
MCM_partition The defined MCM.
}
\description{
Define MCM from a file in which the operators are written as the binary 
representation of the interactions.
}
